Create shift_register.c
This C file can be used to send 8 bit data by using only 3 pins of the microcontroller and is therefore great for applications that require more pins than we have in our controller. Since, sending data to 7-segment LEDs will require atleast 7 pins (not counting .) it would almost use all of the pins of the controller for just 4 7-seg display.